+++ packages/desktop-client/src/components/mobile/transactions/Transaction.jsx
@@ -1,9 +1,13 @@
-import React, { memo, useMemo } from 'react';
+import React, { memo } from 'react';
 
 import { getScheduledAmount } from 'loot-core/src/shared/schedules';
 import { isPreviewId } from 'loot-core/src/shared/transactions';
-import { integerToCurrency, groupById } from 'loot-core/src/shared/util';
+import { integerToCurrency } from 'loot-core/src/shared/util';
 
+import { useAccount } from '../../../hooks/useAccount';
+import { useCategories } from '../../../hooks/useCategories';
+import { usePayee } from '../../../hooks/usePayee';
+import { SvgSplit } from '../../../icons/v0';
 import {
   SvgArrowsSynchronize,
   SvgCheckCircle1,
@@ -41,27 +45,28 @@ ListItem.displayName = 'ListItem';
 
 export const Transaction = memo(function Transaction({
   transaction,
-  account,
-  accounts,
-  categories,
-  payees,
   added,
   onSelect,
   style,
 }) {
-  const accountsById = useMemo(() => groupById(accounts), [accounts]);
-  const payeesById = useMemo(() => groupById(payees), [payees]);
+  const { list: categories } = useCategories();
 
   const {
     id,
     payee: payeeId,
     amount: originalAmount,
     category: categoryId,
+    account: accountId,
     cleared,
     is_parent: isParent,
+    is_child: isChild,
     notes,
     schedule,
   } = transaction;
+
+  const payee = usePayee(payeeId);
+  const account = useAccount(accountId);
+  const transferAcct = useAccount(payee?.transfer_acct);
 
   const isPreview = isPreviewId(id);
   let amount = originalAmount;
@@ -70,10 +75,6 @@ export const Transaction = memo(function Transaction({
   }
 
   const categoryName = lookupName(categories, categoryId);
-
-  const payee = payeesById && payeeId && payeesById[payeeId];
-  const transferAcct =
-    payee && payee.transfer_acct && accountsById[payee.transfer_acct];
 
   const prettyDescription = getDescriptionPretty(
     transaction,
@@ -169,6 +170,15 @@ export const Transaction = memo(function Transaction({
                     color: cleared
                       ? theme.noticeTextLight
                       : theme.pageTextSubdued,
+                    marginRight: 5,
+                  }}
+                />
+              )}
+              {(isParent || isChild) && (
+                <SvgSplit
+                  style={{
+                    width: 12,
+                    height: 12,
                     marginRight: 5,
                   }}
                 />
